requirements.txt
tensorflow
numpy
matplotlib
scikit-learn
seaborn
flask
pillow
werkzeug
joblib
